Moving tests #205
