+++ lisa/lisa-imp/imp-testcases/program-finalization/abstract-class-ext.imp
