+++ lisa/lisa-imp/imp-testcases/program-finalization/default-methods-interface.imp
