+++ lisa/lisa-imp/imp-testcases/program-finalization/signatures-in-concrete-class.imp
