+++ lisa/lisa-imp/imp-testcases/program-finalization/simple-abstract-class.imp
