+++ lisa/lisa-imp/imp-testcases/program-finalization/simple-inheritance.imp
