+++ lisa/lisa-imp/imp-testcases/program-finalization/skip-one.imp
