+++ lisa/lisa-imp/imp-testcases/program-finalization/tree-sanitized.imp
